requirements.txt
adium-theme-ubuntu==0.3.4
Django==1.11.3
django-webpack-loader==0.5.0
djangorestframework==3.6.3
ecdsa==0.13
Fabric==1.10.2
glipper==2.4
gyp==0.1
meld==3.14.2
nose==1.3.7
olefile==0.44
paramiko==1.16.0
pbr==3.1.1
Pillow==4.2.1
psycopg2==2.7.1
pycrypto==2.6.1
pygobject==3.20.0
python-prctl==1.1.1
pytz==2017.2
pyxdg==0.25
six==1.10.0
stevedore==1.24.0
unity-lens-photos==1.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
